Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

So, I get the need for this, but I feel like this is hacking a new feature onto an architecture that isn't quite prepared for it (no offense to the author... it's complicated code). The NodeRequestHandler's run() method is designed to only be called in two scenarios: when a request is first assigned, and on a paused handler. Sprinkling additional run() calls in new places not designed for it I'm almost certain will have bad, unintended consequences (see inline notes).

I'd like for us to consider the above comments as we review. This code is already super fragile, so I'm just trying to exercise an (over) abundance of caution.

Have you had a chance to test this change outside of the unit tests? I'd be curious to see how this performs in a production or devel environment, and under load.

Patch-set: 3
Label: Code-Review=-1