Update patch set 3
Patch Set 3:

I was looking at solving this with the node path that is meant to span nodes, but initially discarded it because it will create a new node in Zookeeper. Thinking about it a little more, this could work by making the static provider aware of those "spawning" nodes which could be used for re-registration of used nodes.

Will look into that, thanks!

Patch-set: 3
Label: Verified=0