Update patch set 3
Patch Set 3:

> How about a job variable that toggles debug (via -d)?  

So this will also foreground the process.  This is incompatible with the extant way of launching which is just 

 - name: Start nodepool launcher
   command: "./.venv/bin/nodepool-launcher -c {{ NODEPOOL_CONFIG }}"

I don't think changing to to " &" is a good idea, as that's not quite daemonizing (might leave behind fd or have other effects on ansible)

However, under the old code, this was using run_process (see https://opendev.org/zuul/nodepool/commit/b5643eadff669369c53026a278898434457790ec) which is devstack's magic to run it as a systemd service.  Running these as services is probably not a bad idea and reflective of reality.
 
Although, I still think we want the logs redirected to files (rather than only captured by the journal) for ease of access.

> We could leave it off for a more production-like setting in nodepool.  Other jobs can set it on either as needed, or, for example, the dib jobs could set it enabled by default.

I'm not sure why we'd turn off debugging in CI jobs, as it is pretty frustrating to get a job failure you don't expect, but then not have the logs to debug it fully.  But I don't mind making it configurable

Patch-set: 3