Disable interface_ip check, when host-key-checking is disable
In the OpenStack driver, we have a check to ensure the interface_ip is
returned by the cloud. This is a way to help confirm there is a public
IP for nodescan to work. However, in some cases we want to disable
host-key-scanning (due to no public IP) which means we also don't want
to fail if the interface_ip is missing.

Change-Id: Ifd2a011302287bffb07b0b94f436748b8620b71d
Signed-off-by: Paul Belanger <pabelanger@redhat.com>