Do not overwrite image upload zk data on delete
This code is setting the ImageUpload state field to DELETING
and updating the data in zookeeper. However, it was using a new
ImageUpload object to do so, and we lose all other data, like the
external_id, external_name, etc. If the delete should fail in the
provider, the next attempt to delete will not have these attributes
(which is not needed to delete, but is used in logging). Keep all
upload attributes so logging will be improved.

Change-Id: I3fbf1c1a3ca10cc397facd4e939cbf4cbb2d8305