DNM: Demonstrate image leak
This test demonstrates the current image leak we have when an
attempt to delete an image upload from a provider fails (e.g.,
because it is associated with a leaked volume that cannot be
deleted because it is associated with a non-existent server).

The bug occurs because we accidentally overwrite many of the ZK
upload record attributes when we change the state to DELETING.
Because we clear out the external_name attribute, if the first
attempt to delete from the provider fails, the next attempt to
delete it will not retry to delete the upload from the provider
and will remove the ZK record.

Change-Id: I391db49ca8db4c6ca22b766aae930deee51f22c2