Update patch set 8
Patch Set 8:

Just discovered that we are missing a crucial piece to dealing with specific image upload records: locking of those records. That should be done here to prevent competing builders from working with the same upload record. There is currently no ZK API for that, so we'll have to add it.

Patch-set: 8