Deregister a launcher when removed from config
We were depending only on the ephemeralness of launcher registration
nodes to indicate when a provider goes away. That only works when
nodepool-launcher is restarted since the ZK connection is shared
among all provider threads. This adds an API method to manually
deregister and modifies a test to ensure it works.

Change-Id: I363fc4a67a83b49b515b963f5d5accdcf1cb758c