Assign static 'building' nodes in cleanup handler
In certain cases there can be a race between a node being re-registered
and a request handler creating a new 'building' node. This will lead to
a starvation of the 'building' node until a node with a matching label
is re-registered again (if at all).

To avoid this, the cleanup handler will check if there are any ready
nodes that can be assigned to nodes waiting for a matching label.

Change-Id: Ifc7fe2c59d5c4393c462f508ed449736ae145371