Update patch set 2
Patch Set 2:

The problem is, that the node should not be gone. It's easy to write a test cases that passes with the bug in place but fails with the fix. But for a regression test we need it the other way around and that seems to have a race.

Patch-set: 2
Label: Verified=0