Fix builder shutdown race in tests
The builder intentionally does not attempt to shutdown the uploader
threads since that could take an unreasonable amount of time. This
causes a race in our tests where we can shutdown the ZooKeeper
connection while the upload thread is still in progress, which will
cause the test to fail with a ZooKeeper error. This adds a test
cleanup for the builder used in tests.

Change-Id: I25d4b52e17501e5dc6543adef585dd3b86bd70f9