WIP: Implement a Devnest nodepool driver
Devnest is a simple CLI to manage hardware "reservations".

Currently devnest is implemented to interact with Jenkins server
and allows to take node from jenkins and assign it to be run as
zuul executor. At the same time hardware can be used reserved
manually by users.

It is available via PyPI or as github project:

    https://github.com/rhos-infra/devnest.git

Depends-On: https://review.opendev.org/#/c/639632/

Co-authored-by: Tristan Cacqueray <tdecacqu@redhat.com>
Change-Id: I2c06d68782e63f9a8dee1e7dfbfbce0385a40dcf