Dockerfile: install nodepool-builder dependencies
vhd-util is needed for targetting Rackspace. debian-keyring is
needed for debian images, ubuntu-keyring for ubuntu images and
yum for red hat images.

Note the sibling build will have installed many of these from the
bindep.txt file from diskimage-builder itself.  However, when using
releases this is not done.  These installs should be a no-op for the
sibling containers.

Change-Id: I35bc6a2a07fda229acfd53a2a34227d6475495a8