Update patch set 1
Patch Set 1: Code-Review+2

I gotta be honest; this is based on buster [1] and mixing in packages built for bionic, and is an environment we don't have validated by any CI builds which are all based on bionic.  Buster rpm does seem to build against libzstd [2] which is at least the immediate issue for fedora builds; what is the guarantee from the underlying python-slim [3] about the distribution and what might change?

I can feel I'm on the loosing team so this is certainly closer to being what is necessary, so gets my vote.  

[1] https://zuul.opendev.org/t/zuul/build/091e8cbc8d254797b9c7fe1c8c9a42af/log/job-output.txt#2863
[2] https://metadata.ftp-master.debian.org/changelogs//main/r/rpm/rpm_4.14.2.1+dfsg1-1_changelog
[3] https://opendev.org/opendev/system-config/src/branch/master/docker/python-builder/Dockerfile#L17

Patch-set: 1
Label: Code-Review=+2