Delete images by id
Nodepool currently deletes managed images by name. This leads to
problems when an image is uploaded twice (e.g. because of multiple
formats). In this case there can be more than one image with the same
name which breaks the deletion. This can be fixed by deleting the
images by id instead.

Change-Id: I74fc6219ef7f2c496f36defb0703137ec4d7d30e