Update patch set 2
Patch Set 2:

recheck

i have tested this manually and it lets dib get started.  there is an bug within dib ... it's looking for an init system on the host, which the container doesn't have.  this is leaking; it should be looking in the chroot, not the host.  i'm sure we've debugged something similar before; will look into it.

Patch-set: 2