Update patch set 2
Patch Set 2: Code-Review-1

Although nothing is outwardly broken by this change, it does affect the fake driver since the provider.py code does not expect a dict passed in to the FakeOpenStackCloud.delete_image() method there (name_or_id is checked in the subsequent _delete() call).

That being said, that fake driver code seems inherently broken since the images are not showing up in the self._image_list, which is something else we should investigate. But I think we can go ahead and handle the dict passed in, then fix the other thing.

Patch-set: 2
Label: Code-Review=-1