Make opendevorg/ namespace docker images
This creates opendevorg/ namespace docker images.  They are based on
the dib and openstacksdk opendev images, which includes everything to
build images for OpenDev providers.

This will be followed-on by testing the containerised nodepool-builder
images in the devstack based functional test (i.e. building and then
uploading to/booting in the devstack cloud)

Depends-On: https://review.opendev.org/693971
Depends-On: https://review.opendev.org/696015
Change-Id: Ifec8c8d078eec1946c29d1ea49414650111ad3ac