docker: Base nodepool images on diskimage-builder base image
This bases the nodepool images on an image with diskimage-builder and
dependencies installed.

This is an alternative to https://review.opendev.org/693306

This will be followed-on by testing the containerised nodepool-builder
images in the devstack based functional test (i.e. building and then
uploading to/booting in the devstack cloud)

Depends-On: https://review.opendev.org/694175
Change-Id: Ifec8c8d078eec1946c29d1ea49414650111ad3ac