Make zuul/ namespace docker images
This migrates the existing single dockerfile to Dockerfile.zuul

The zuul/ namespace images will be the reference implementation.  It
is based on released version of tools (openstacksdk) and the builder
does not include dib dependencies (users could further customise this,
if they are indeed even using dib).

This is supposed to be considered in the context of the follow-on
Ifec8c8d078eec1946c29d1ea49414650111ad3ac which creates different
OpenDev specific images in the opendev/ namespace.  There we choose to
use master branches of tools and bundle in everything to make dib
images suitable for the OpenDev infra providers.

Change-Id: Id608a00b5d4bd90290a3d90a04e2cd21186a276a