WIP: decouple provider manager from config objects
The ProviderManager currently stores the Provider objects within
the config structure created and used by the builder or launcher process.
This is an artifact of the original nodepool, and is one of the most
confusing parts of the base nodepool structure. This experiments with
making the ProviderManager a singleton that maintains its own state.

Change-Id: I1744356936c7c633aefb6c916e9b347dc4294ad5