Update patch set 1
Abandoned

Turns out a singleton is not really a good fit for this. Since we have multiple threads that each call ProviderManager.reconfig() on a config file change, we'd need to use a lock (not that big of a deal), but also not reconfigure more than needed (i.e., only once). I think that last bit voids this idea.

Patch-set: 1
Status: abandoned