Add back-off mechanism for image deletions
Some ceph based clouds occationally leak volumes of instances when the
instance is getting deleted. This blocks the cleanup of the image
because a snapshot blocks the deletion. On deployments with many
builders and a growing number of non-deletable images this leads to
hammering the glance service with delete requests.

A solution can be to first coordinate the deletion of the images
combined with a back-off mechanism.

Change-Id: I1b4aecd6ee200bd179f3cab24aa43a308bb86319