Always identify static nodes by node tuple
In some places in the static provider the node was identified by
hostname instead of the node tuple (hostname, username, port).

In case a node is registered with the same multiple times using
different users and/or ports this caused problems in the code paths that
were expecting the hostname to be unique.

This was the case in the liveness check where the provider e.g. did not
check different ports. Also the re-registration of deleted nodes was not
using the correct node config in this case.

Log messages were modified to log the node tuple instead of the
hostname.

Change-Id: I7ead7e9e6903249449d350dc35deffaf98171304