Cleanup exception logging in static provider
The static provider was logging a lot of exceptions with traceback in
case of an unreachable nodes.

Since offline nodes are part of the normal operation (e.g. or
maintenance) and we don't gain any additional insights from the
traceback, the log level was reduced to warning on StaticNodeErrors.

The warning messages will still log the exception message, so we don't
loose any important information. In case of other exceptions we will
still print the full traceback.

Change-Id: I5a6035a090acac2edb8529e91be63995feceb23d