Centralize logging adapters
Logging adapters were already in use for the NodeLauncher. In order to
expand logging of Zuul's event id to nodepool, we will first centralize
the existing adapters and refactor them simliar to how this is already
done in Zuul.

This will allow us to correlate node request with other log messages
from Zuul via the event id to allow for faster and easier debugging.

Change-Id: Idc7468239a055c75074e876ea8c445b49c791131