Pass node request handler to launcher base class
All implementations of the NodeLauncher have access to the node request
handler instance, but it was not passed through to the base class.

In order to initialize the event log adapter, the NodeLauncher will now
receive then node request handler instead of the Zookeeper connection.

The Zookeeper connection is directly extracted from the request handler.

Change-Id: I8ee7335f50b64435c82a5f039989f98750874466