Update patch set 1
Patch Set 1:

The check version of the job should be nodepool-build-image the gate version is nodepool-upload-image because it uploads the images also to docker hub (which we don't want to do in check). Recheck has been done to check if it's a transient error (I've seen this job failing in the past for transient reasons).

Patch-set: 1