Lower min-ready request priority
When at capacity min-ready requests are processed with high priority
("100"). Then if that node type isn't needed currently it will be deleted
to make room for new requests. Unfortunately because the next min-ready
request has a high priority it is handled again. This then runs in a
loop preventing allocation of resoures to node labels that need
resources.

Address this by reducing the priority for min-ready requests to "900" so
that they will only be processed when job demand is low.

Note that request priorities are sorted as strings. From high to low
priority:

  "100"
  "1000"
  "200"
  "900"

Change-Id: If8cddc418da62a135abc136d7b9a87f6a3073c46