Update patch set 1
Patch Set 1: Code-Review-1

Ok, so here's the deal on this (after I've had a chance to properly think about it). We aren't actually in control of the node ID creation: ZK is. It creates them as 0-leading numerical strings automatically (a ZK sequence ID) when we create a new node. For us to transform those into integers for our purposes means we'd have to transform them BACK into the same 0-prepended string to update existing nodes (which we do a lot). This opens us up for bugs. True, we could contain that logic in only a couple of places, but there is nothing being fixed here, really. I say if it ain't broke, don't fix it, and let's leave it be. I'm therefore against this change.

Patch-set: 1
Reviewer: Gerrit User 3099 <3099@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1