k8s: allow port forwarding
Allow the user created for Zuul to configure pod port forwarding.
This is in service of supporting log streaming from pods.  This
may also allow users to listen on arbitrary ports on the executor;
if that is not desired, we can create a second set of credentials
for Zuul itself and configure RBAC for that, so that Zuul can
perform this action but users may not.

Change-Id: I17a28070a4cfd40bf09dcd0baa4d293b49b14651