Update patch set 1
Patch Set 1:

I think this is okay -- I don't think this can be used nefariously (untrusted users won't be able to run "kubectl port-forward" on the executor (and even if it did, they could only listen on a high port anyway); even if they can configure port-forwarding on the k8s side, that doesn't affect the executor; and I don't think this allows users to escape the k8s namespace).  But I'd love more eyes on this in case it could cause a problem I didn't anticipate.

Patch-set: 1
Label: Verified=0