Update patch set 32
Patch Set 32:

(4 comments)

Some minor thoughts inline.

It also appears that the testing is not using SSL. Is that something we should add to the testing as this is how we expect people to use authenticated access in the wild?

Patch-set: 32
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0