Kubernetes/OpenShift Provider: Don't Require Bash in Container Images
Currently the Kubernetes and OpenShift providers set the entrypoint
of their build node pods to `/bin/bash`, which then requires `bash`
to be available in the respective container image. This might not
always be the case (e.g. with Alpine based images).
This change makes sure the entrypoint is set to `/bin/sh`, which we
can more reliably assume to be available in the container image.

Change-Id: I799ea95b715e50d9c22e66cc80579cf119db8f38