Stop using hacking for anything
We've had it in here as a way to pin pycodestyle. But that's not
worth the energy to deal with things like hacking now spewing
incorrect messages about our apache license headers not being
right even though they are. (it's because sdague dedcided that
nobody has more than 11 lines of header)

zuul just uses a bare flake8. Match that choice.

Change-Id: I5bc847d151801fd31281389ee65184e964ddd2a5
Depends-On: https://review.opendev.org/712778