Add visual dividers for each image in Dockerfile
Our dockerfiles describe the building of many images. Some are temporary
and others persist to publishing, but they all have specific roles and
knowing where to apply changes is important. Unfortanately, it is hard
to see that by default. Experiment with the addition of ==== barrier
lines to give people a visual cue for the divisions.

Change-Id: If28e0bd94a4d65f0623c56e9a589615dd04d2e75