Add diskimages-globals section for common configuration
A new config section diskimages-globals is added to allow common
values of elements and env-vars to be set for all diskimages at once.

While YAML does have inbuilt support for anchors to greatly reduce
duplicated sections, anchors have no support for merging values.  For
diskimages, this can result in a lot of duplicated values for each
image which you can not otherwise avoid.

While I'm fairly cognisant we don't want to turn the configuration
files into a badly-done bespoke templating language, I think this has
good potential to greatly reduce config file verbosity in many cases.

Documentation is updated, and a test is added to ensure the correct
merging and override behaviour of the new section.

Change-Id: I170016ef7d8443b9830912b9b0667370e6afcde7