Dockerfile: set SECLEVEL=1 for RAX dfw and ord region compatability
SSL connections are refused in the container environment to
<dfw|ord>.images.api.rackspacecloud.com (*not* iad, note) due to SSL
errors.

In python it manifests as something like

 [SSL: WRONG_SIGNATURE_TYPE] wrong signature type (_ssl.c:1076)

but you can easily replicate it with a curl too where you get
something else like

 curl: (35) error:1414D172:SSL routines:tls12_check_peer_sigalg:wrong
 signature type

Emperical evidence shows that when turning the SECLEVEL to 1 we see in
the connection

 Peer signing digest: SHA1

which is now considered insecure [1].  *Why* we get this is unclear,
and seems that it must have to do wish misconfiguration on the RAX API
side.

Workaround this for now by turning the SECLEVEL down.

[1] https://wiki.debian.org/ContinuousIntegration/TriagingTips/openssl-1.1.1

Change-Id: I932445d21881c487086af3110f6e870a2d70bfa8
Story: #2007407
Task: #39073