Update patch set 2
Patch Set 2:

> IIUC this may be problematic when some images need a specific non-global configuration, for example a tenant disk with a custom key or something.

I'm not quite following this, if it's not something that applies to all your images, don't put it in the global and everything is the same?

> Perhaps it would be easier to model the diskimages configuration using parent inheritance like zuul does with jobs?

We could make an entry in the diskimages section have an abstract:true value and then if another entry has parent: that it can merge from there.  But I only think it makes sense for env-vars and elements fields still because they can't have defaults.

An abstract could use another abstract as a parent and keep merging down ... I'm unsure if this is getting unnecessarily complex for "I want this key in env for all images".

Are we at least ok with the way this is merging?  If this happens via abstract entries or like this is the same, and if that's not acceptable both ideas are non-starters...

Patch-set: 2
Label: Verified=0