Update patch set 2
Patch Set 2:

> > IIUC this may be problematic when some images need a specific
 > non-global configuration, for example a tenant disk with a custom
 > key or something.
 > 
 > I'm not quite following this, if it's not something that applies to
 > all your images, don't put it in the global and everything is the
 > same?
 > 

I meant that this doesn't work if you want to apply a variable to a
subset of all the diskimages. For example it would be nice to also
centralize the elements needed by fedora 30, 31 and rawhide diskimage.
Using parent inheritance would help solve the same problem once.

Patch-set: 2
Label: Verified=0