DiskImage : alphabetise entries
This is a no-op refactor to just keep the fields in alphabetical
order.  This makes it much easier to visually keep the initalizer and
comparision functions in sync across any future changes that may
update the fields.

Note name is kept first, out of order, to signal it is a primary key
(see next follow-on I3931dc1457c023154cde0df2bb7b0a41cc6f20d3 where
this is enforced)

Change-Id: I0c177b3b0edee3cd811c029d3f03a08c9e11a5fb