Update patch set 4
Patch Set 4:

I think this is self-signalling that it's not a good idea.  I was using __slots__ as sort of a poor version of static typing (which would be appropriate for this object) but it's not necessary.

I will just alphabetize the extant fields, which makes it much easier to compare the initalizier and comparison functions if and when fields get updated.

Patch-set: 4
Label: Verified=0