Add libc6-dev to bindep
The python-builder base image was updated to no longer install
recommends. This is inline with the other Infra images and keeps
image sizes smaller. gcc recommended libc6-dev - but it turns out
we need that for limits.h for one of our depends. Add it to fix
our image builds.

Change-Id: Ic7db84155f8d5f61007dfe47e380dccade0992cc