Fix shutdown ordering
The taskmanager driver did not stop the taskmanager thread
correctly.  It is updated to do so.

The launcher shutdown order stopped providers before the cleanup
threads, however, cleanup threads may be in the middle of interacting
with providers, which can cause them to hang.  Instead, stop the
cleanup threads first, then the providers.  Without this, failing
unit tests may not exit correctly, making debugging difficult.

Change-Id: I06256ddca38b815bf61ae3f710e000d666c3ef4b