k8s/OKD Provider: Don't Set ca_cert if TLS verification is skipped
Kubernetes does not allow to set a ca_cert in a kubeconfig if TLS
certificate verifiaction is disabled. Doing so results in an error
message:
`error: specifying a root certificates file with the insecure flag is not allowed`
This change makes sure we skip the ca_cert option nodepool-launcher
generates for the Zuul executor if nodepools kubeconfig is set to
skip TLS cert verification.

Change-Id: I458c054fc9fae340d187ce40ea1236efdf65d50f