Update patch set 1
Patch Set 1: Code-Review+2

Thanks for this.

My first thought was "surely the unit tests must cover this", but it turns out we only ever check the default.  Since I modified a test to see if that was right, i'll propose it as an follow-on rather than request this be updated -- but in general it would be better to include a unit test for something like this.

Patch-set: 1
Label: Code-Review=+2