Add Cobbler driver
The Cobbler driver is modelled after the Kubernetes driver with label
type 'namespace'.  The driver connects to a Cobbler server and query the
server for an inventory of machines that the driver can provision.

When the driver provision a machine/node, the driver will return a
coordinate that you can use to manage the allocated node via the
'zuul.resources.<node name>.namespace' variable.  The value of the
variable is a base64-encoded JSON string.

Similar to the 'namespace' type of Kubernetes driver, the returned
resource needs to be setup before it can be use.  The 'bmc' value is the
value of 'Power Management Address' value of the system in Cobbler.  You
would manage the allocated node with the 'bmc' value using protocols like
IPMI and connect to the node with the 'node' value using protocols
like SSH.

Testing the Cobbler driver:
tox -e py3 -- test_cobbler_config
tox -e py3 -- test_cobbler_basic

Change-Id: Ib875064475d6b44b2861a81cabad62920eee71dc