Run upload-docker-image on release
When we tag a release of Nodepool, run the upload-docker-image job, but
tell it to do so in a non-promote mode.  This will cause it to
directly upload the images to Docker Hub with the final tags applied.
We want to do this in the release pipeline so that we get images
built with the correct version number based on the git tag.

When run in a pipeline with a tag attribute set, this will use all
three lengths of the version number (3.19.0, 3.19, 3) as tags instead
of the single 'latest' tag which is what is applied in the promote
pipeline.

This temporarily comments out the other release jobs so that we can
manually enqueue the most recent tag and retroactively build and
publish it to Docker Hub.  Once this works, we will re-enable those
two jobs.

Apparently we also need to use python3 in this job now.  We were
already doing that for Zuul.  The default will be switching soon,
and we can remove that var entry then.

Change-Id: Id85bc3f56a9c18f1e6f4b14e6edefe76eabeff3e